feat: create upgrade workflow for cdklabs-projen-project-types
This PR creates a new workflow which will apply to all managed project
types. The workflow will control upgrade the
`cdklabs-projen-project-types` dependency (this project). The goal is to
have a workflow that has no other dependencies and can be executed as a
stand alone workflow. This will allow us to push out changes to all of
our projects from a central location.

A follow up to this will be to add a script in `cdk-ops` manually
trigger this workflow across all of the repos.