fix(monorepo): repository url is not passed to workspaces
Previously, workspaces couldn't inherit the URL of the parent monorepo
because it wasn't exposed. This allows setting the URL just once on the
monorepo and having it applied to all workspaces (where it can still be
overridden). Since the construct is for a monorepo, this should be a
sensible default.