fix: decouple `workflowNodeVersion` from `minNodeVersion`
`workflowNodeVersion` is copied from `minNodeVersion` if not set.

This incentivizes undesirable behavior: `minNodeVersion` should almost
never need to be set, while `workflowNodeVersion` might need to be set.

As a result, users might set `minNodeVersion` while looking for the
other version.

Fix that in the types that wrap projen types here, in preparation of
an upstream projen fix.