fix: decouple `workflowNodeVersion` from `minNodeVersion` (#610)
* fix: decouple `workflowNodeVersion` from `minNodeVersion`

`workflowNodeVersion` is copied from `minNodeVersion` if not set.

This incentivizes undesirable behavior: `minNodeVersion` should almost
never need to be set, while `workflowNodeVersion` might need to be set.

As a result, users might set `minNodeVersion` while looking for the
other version.

Fix that in the types that wrap projen types here, in preparation of
an upstream projen fix.

* chore: self mutation

Signed-off-by: github-actions <github-actions@github.com>

* Update documentation for copied interfaces

* Also do the same for the builtin `CdkConstructLibrary` class

* Fix build

* chore: self mutation

Signed-off-by: github-actions <github-actions@github.com>

---------

Signed-off-by: github-actions <github-actions@github.com>
Co-authored-by: github-actions <github-actions@github.com>