fix: versioning options are not respected in a monorepo
Versioning options for releasing did not use to be supported, because
they are being passed in a funky way in upstream projen code.

They need to be part of the *invocation* of the `bump` task instead of
the `bump` task itself, since we might need multiple invocations for
multiple configured release branches (and there can be only one task
with the name `bump`).

Do the same in the monorepo.