fix: versioning options are not respected in a monorepo (#726)
* fix: versioning options are not respected in a monorepo

Versioning options for releasing did not use to be supported, because
they are being passed in a funky way in upstream projen code.

They need to be part of the *invocation* of the `bump` task instead of
the `bump` task itself, since we might need multiple invocations for
multiple configured release branches (and there can be only one task
with the name `bump`).

Do the same in the monorepo.

* chore: self mutation

Signed-off-by: github-actions <github-actions@github.com>

* Finish thought in comment

---------

Signed-off-by: github-actions <github-actions@github.com>
Co-authored-by: github-actions <github-actions@github.com>