fix(cli-integ-tests): fetch tags from upstream
When the CLI integ tests run from a fork, the fork won't have the
version tags and the candidate bumping fails (everything gets bumped
to `0.0.0`, which causes tests to fail).

Instead, fetch tags from upstream.

I've opted to make the repo to fetch from an explicit parameter; I
could have guessed that the repo would be `aws/<fork-name>`, but
that just seemed like another avenue to get subtle bugs in the future.

Since this is now a newly required parameter which is breaking, I
might as well remove the parameter that was only temporarily there
for the duration of the migration.