fix: gather-versions doesn't work for 0 or 1 dependencies
When I simplified the command-line that `gather-versions` takes, I
forgot to take out the condition that requires a minimum number of
arguments.

The result is that `gather-versions` won't do anything thing for
packages that have 0 or 1 workspace dependency; plus, the script didn't
actually exit with an error code when that happened, so it was hard to
detect.

Fix that in this PR, and also add a test that runs the `gather-versions`
task properly from a projen monorepo config using `TaskRuntime`.