fix: gather-versions doesn't work for >= 2 dependencies  (#833)
* fix: gather-versions doesn't work for 0 or 1 dependencies

When I simplified the command-line that `gather-versions` takes, I
forgot to take out the condition that requires a minimum number of
arguments.

The result is that `gather-versions` won't do anything thing for
packages that have 0 or 1 workspace dependency; plus, the script didn't
actually exit with an error code when that happened, so it was hard to
detect.

Fix that in this PR, and also add a test that runs the `gather-versions`
task properly from a projen monorepo config using `TaskRuntime`.

* fix: gather-versions doesn't work for >= 2 dependencies

During refactoring I accidentally removed a `.join(' ')` so an
array stringifies fine but stringifies to the wrong string. Instead of:

```
a b c
```

it stringifies to

```
a,b,c
```

Which is not a command-line argument per package, but once command-line
argument with all packages in there.