feat: add support for selective releases
Add a user interface option to do selective releasing in the GitHub
workflow.

This adds a string input that should hold a JSON array with the
packages to release. If the array is empty (or the input is not
present), then we release all packages.

We do the build for all packages, but hook into the "needs-publish"
step to add this condition into that expression:

```yaml
    outputs:
      publish-xyz: ${{ steps.check-publish-xyz.outputs.publish }}
```

to

```yaml
    outputs:
      publish-xyz: ${{ steps.check-xyzy-schema.outputs.publish && (fromJSON(inputs.package_list || '[]')[0] == null || contains(fromJSON(inputs.package_list || '[]'), 'xyz')).outputs.publish }}
```

I'm not sure how to best test this, except doing it for real in
a low-risk repo (like the service spec repo).