feat: beginnings of RetryAutoMerge (#881)
At some point this will turn into a component that will automatically
switch AutoMerge back on if it gets switched off due to a build/test
workflow sporadically failing in the merge queue.

(Note that for automerge to be switched off, the workflow would have
first needed to succeed on the PR, and then subsequently fail in the
merge queue)

We don't know how to properly filter for that event yet, so we're
just logging the entire event contents to look at later if it happens
again.